pyproject.toml
[project]
name = "minichatagent"
version = "0.5.0"
description = "MiniChatAgent: an AI bot framework powered by NoneBot2 (based on Amrita)"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10, <4.0"
dependencies = [
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "click>=8.2.1",
    "colorama>=0.4",
    "toml>=0.10.2",
    "pip>=25.2",
    "nb-cli>=1.0.0",
    "nonebot-plugin-localstore>=0.7.4",
    "typing-extensions>=4.6.0",
    "uv>=0.8.12",
    "requests>=2.0", # See https://github.com/twindb/backup/pull/465
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "packaging>=25.0",
    "zipp>=3.23.0",
    "pytz>=2025.2",
]

[project.optional-dependencies]
full = [
    "aiohttp>=3.13.2",
    "pillow>=12.0.0",
    "fastmcp>=2.13.0.2",
    "bcrypt>=4.3.0",
    "async-lru>=2.0.5",
    "jinja2>=3.1.6",
    "uvicorn>=0.35.0",
    "psutil>=7.0.0",
    "pytz>=2025.1",
    "stubs>=1.0.0",
    "dotenv>=0.9.9",
    "importlib>=1.0.4",
    "watchfiles<1.0.0",
    "openai>=1.66.3, <=1.99.0",
    "pydantic>=2.4.2",
    "jieba>=0.42.1",
    "nonebot-plugin-orm[default]>=0.8.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.3",
]

[project.scripts]
minichatagent = "amrita.cli:main"
miniagent = "amrita.cli:main" # backward-compatible alias
amrita = "amrita.cli:main" # backward-compatible alias

[project.urls]
"Homepage" = "https://github.com/momo-mnsjtxy/MiniAgent"
"Source" = "https://github.com/momo-mnsjtxy/MiniAgent"
"Issue Tracker" = "https://github.com/momo-mnsjtxy/MiniAgent/issues"
"Upstream" = "https://github.com/LiteSuggarDEV/Amrita"

[tool.setuptools.packages.find]
include = ["amrita", "amrita.*"]

# 添加资源文件包含配置
[tool.setuptools.package-data]
"amrita.plugins.webui.service" = ["templates/*.html", "static/**/*"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.8",
    "nonebot-plugin-orm[default]>=0.8.2",
    "pytest>=8.4.1",
    "minichatagent[full]",
    "mcp>=1.21.1",
    "pyright>=1.1.407",
]
package = true

[tool.uv.sources]
minichatagent = { workspace = true }

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"

[tool.amrita]
plugins = [
    "amrita.plugins.chat",
    "amrita.plugins.manager",
    "amrita.plugins.webui",
    "amrita.plugins.search",
    "amrita.plugins.linuxcmd",
    "amrita.plugins.codegen",
]

[tool.nonebot]
plugins = ["nonebot_plugin_orm"]
